Uniform crossover was all sorts of broken: it assumed genesLoci was missing at least one extreme, it was only randoming 64 genes and even those it was doing incorrectly. Hopefully everything is fixed now

+++ include/genetic-algorithm.hpp
@@ -9,12 +9,12 @@ void exponentialRanking(int populationSize, float *fitness, bool maximizeFitness
 
 void tournamentRanking(int populationSize, float *fitness, bool maximizeFitness, int tournamentSize, int *winners, int winnersSize);
 
-void twoPointsCrossover(uint8_t *parent1, uint8_t *parent2, int length, uint8_t *child, uint64_t *genesLoci, int genesLociLength);
+void twoPointsCrossover(uint8_t *parent1, uint8_t *parent2, uint64_t length, uint8_t *child, uint64_t *genesLoci, int genesLociLength);
 
-void twoPointsCrossover(uint8_t *parent1, uint8_t *parent2, int length, uint8_t *child);
+void twoPointsCrossover(uint8_t *parent1, uint8_t *parent2, uint64_t length, uint8_t *child);
 
-void uniformCrossover(uint8_t *parent1, uint8_t *parent2, int length, uint8_t *child);
+void uniformCrossover(uint8_t *parent1, uint8_t *parent2, uint64_t length, uint8_t *child);
 
-void uniformCrossover(uint8_t *parent1, uint8_t *parent2, int length, uint8_t *child, uint64_t *genesLoci, int genesLociLength);
+void uniformCrossover(uint8_t *parent1, uint8_t *parent2, uint64_t length, uint8_t *child, uint64_t *genesLoci, int genesLociLength);
 
 void mutate(uint8_t *individual, int length, float mutationProbability);
